Refactor as part of online test pass mark report to support video interview scores
