+++ app/controllers/ReportingController.scala
@@ -24,13 +24,14 @@ import model.report.{ DiversityReportItem, OnlineTestPassMarkReportItem, Progres
 import play.api.libs.json.Json
 import play.api.mvc.{ Action, AnyContent, Request }
 import repositories.application.GeneralApplicationRepository
+import repositories.application.ReportingRepository
 import repositories.{ QuestionnaireRepository, _ }
 import uk.gov.hmrc.play.microservice.controller.BaseController
 
 import scala.concurrent.ExecutionContext.Implicits.global
 
 object ReportingController extends ReportingController {
-  val appRepository = applicationRepository
+  val reportRepository = reportingRepository
   val cdRepository = contactDetailsRepository
   val authProviderClient = AuthProviderClient
   val questionnaireRepository = repositories.questionnaireRepository
@@ -43,7 +44,7 @@ trait ReportingController extends BaseController {
 
   import Implicits._
 
-  val appRepository: GeneralApplicationRepository
+  val reportRepository: ReportingRepository
   val cdRepository: ContactDetailsRepository
   val authProviderClient: AuthProviderClient
   val questionnaireRepository: QuestionnaireRepository
@@ -54,7 +55,7 @@ trait ReportingController extends BaseController {
   def adjustmentReport(frameworkId: String) = Action.async { implicit request =>
     val reports =
       for {
-        applications <- appRepository.adjustmentReport(frameworkId)
+        applications <- reportRepository.adjustmentReport(frameworkId)
         allCandidates <- cdRepository.findAll
         candidates = allCandidates.groupBy(_.userId).mapValues(_.head)
       } yield {
@@ -72,12 +73,12 @@ trait ReportingController extends BaseController {
   }
 
   def candidateProgressReport(frameworkId: String) = Action.async { implicit request =>
-    appRepository.candidateProgressReport(frameworkId).map(r => Ok(Json.toJson(r)))
+    reportRepository.candidateProgressReport(frameworkId).map(r => Ok(Json.toJson(r)))
   }
 
   def diversityReport(frameworkId: String) = Action.async { implicit request =>
     val reports = for {
-      applications <- appRepository.diversityReport(frameworkId)
+      applications <- reportRepository.diversityReport(frameworkId)
       questionnaires <- questionnaireRepository.findAllForDiversityReport
       medias <- medRepository.findAll()
     } yield {
@@ -96,7 +97,7 @@ trait ReportingController extends BaseController {
   def onlineTestPassMarkReport(frameworkId: String) = Action.async { implicit request =>
     val reports =
       for {
-        applications <- appRepository.onlineTestPassMarkReport(frameworkId)
+        applications <- reportRepository.onlineTestPassMarkReport(frameworkId)
         questionnaires <- questionnaireRepository.findForOnlineTestPassMarkReport
       } yield {
         for {
@@ -116,14 +117,14 @@ trait ReportingController extends BaseController {
     preferencesAndContactReports(nonSubmittedOnly = false)(frameworkId)
 
   def applicationAndUserIdsReport(frameworkId: String) = Action.async { implicit request =>
-    appRepository.allApplicationAndUserIds(frameworkId).map { list =>
+    reportRepository.allApplicationAndUserIds(frameworkId).map { list =>
       Ok(Json.toJson(list))
     }
   }
 
   private def preferencesAndContactReports(nonSubmittedOnly: Boolean)(frameworkId: String) = Action.async { implicit request =>
     for {
-      applications <- appRepository.applicationsReport(frameworkId)
+      applications <- reportingRepository.applicationsReport(frameworkId)
       applicationsToExclude = getApplicationsNotToIncludeInReport(applications, nonSubmittedOnly)
       users <- getAppsFromAuthProvider(applicationsToExclude)
       contactDetails <- cdRepository.findAll
