+++ app/repositories/application/GeneralApplicationRepoBSONToModelHelper.scala
@@ -16,234 +16,18 @@
 
 package repositories.application
 
-import connectors.launchpadgateway.exchangeobjects.in.reviewed.{ ReviewSectionQuestionRequest, ReviewedCallbackRequest }
 import model.ApplicationStatus.{ apply => _, _ }
 import model.AssessmentScheduleCommands.ApplicationForAssessmentAllocation
-import model.CivilServiceExperienceType.{ CivilServiceExperienceType, apply => _, _ }
-import model.Commands._
-import model.InternshipType.{ InternshipType, apply => _ }
-import model.OnlineTestCommands.TestResult
-import model.SchemeType._
-import model.command.ProgressResponse
+import model.CivilServiceExperienceType.{ apply => _ }
+import model.InternshipType.{ apply => _ }
 import model.persisted._
-import model.report._
-import model.{ CivilServiceExperienceType, InternshipType }
 import org.joda.time.DateTime
-import play.api.Logger
 import reactivemongo.bson.{ BSONDocument, _ }
 import repositories._
 
 object GeneralApplicationRepoBSONToModelHelper extends GeneralApplicationRepoBSONToModelHelper
 
-trait GeneralApplicationRepoBSONToModelHelper {
-
-  def toReportWithPersonalDetails(findProgress: (BSONDocument, String) => ProgressResponse)
-                                 (doc: BSONDocument): ReportWithPersonalDetails = {
-    val fr = doc.getAs[BSONDocument]("framework-preferences")
-    val fr1 = fr.flatMap(_.getAs[BSONDocument]("firstLocation"))
-    val fr2 = fr.flatMap(_.getAs[BSONDocument]("secondLocation"))
-
-    def frLocation(root: Option[BSONDocument]) = extract("location")(root)
-    def frScheme1(root: Option[BSONDocument]) = extract("firstFramework")(root)
-    def frScheme2(root: Option[BSONDocument]) = extract("secondFramework")(root)
-
-    val personalDetails = doc.getAs[BSONDocument]("personal-details")
-    val aLevel = personalDetails.flatMap(_.getAs[Boolean]("aLevel").map(booleanTranslator))
-    val stemLevel = personalDetails.flatMap(_.getAs[Boolean]("stemLevel").map(booleanTranslator))
-    val firstName = personalDetails.flatMap(_.getAs[String]("firstName"))
-    val lastName = personalDetails.flatMap(_.getAs[String]("lastName"))
-    val preferredName = personalDetails.flatMap(_.getAs[String]("preferredName"))
-    val dateOfBirth = personalDetails.flatMap(_.getAs[String]("dateOfBirth"))
-
-    val fpAlternatives = fr.flatMap(_.getAs[BSONDocument]("alternatives"))
-    val location = fpAlternatives.flatMap(_.getAs[Boolean]("location").map(booleanTranslator))
-    val framework = fpAlternatives.flatMap(_.getAs[Boolean]("framework").map(booleanTranslator))
-
-    val ad = doc.getAs[BSONDocument]("assistance-details")
-    val needsAssistance = ad.flatMap(_.getAs[String]("needsAssistance"))
-    val needsAdjustment = ad.flatMap(_.getAs[String]("needsAdjustment"))
-    val guaranteedInterview = ad.flatMap(_.getAs[String]("guaranteedInterview"))
-
-    val applicationId = doc.getAs[String]("applicationId").getOrElse("")
-    val userId = doc.getAs[String]("userId").getOrElse("")
-    val progress: ProgressResponse = findProgress(doc, applicationId)
-
-    val onlineTests = doc.getAs[BSONDocument]("online-tests")
-    val cubiksUserId = onlineTests.flatMap(_.getAs[Int]("cubiksUserId"))
-
-    ReportWithPersonalDetails(
-      applicationId, userId, Some(ProgressStatusesReportLabels.progressStatusNameInReports(progress)),
-      frLocation(fr1), frScheme1(fr1), frScheme2(fr1),
-      frLocation(fr2), frScheme1(fr2), frScheme2(fr2), aLevel,
-      stemLevel, location, framework, needsAssistance, needsAdjustment, guaranteedInterview, firstName, lastName,
-      preferredName, dateOfBirth, cubiksUserId
-    )
-  }
-
-  def toCandidateProgressReportItem(findProgress: (BSONDocument, String) => ProgressResponse)
-                                   (doc: BSONDocument): CandidateProgressReportItem = {
-    val schemesDoc = doc.getAs[BSONDocument]("scheme-preferences")
-    val schemes = schemesDoc.flatMap(_.getAs[List[SchemeType]]("schemes"))
-
-    val adDoc = doc.getAs[BSONDocument]("assistance-details")
-    val disability = adDoc.flatMap(_.getAs[String]("hasDisability"))
-    val onlineAdjustments = adDoc.flatMap(_.getAs[Boolean]("needsSupportForOnlineAssessment")).map(booleanTranslator)
-    val assessmentCentreAdjustments = adDoc.flatMap(_.getAs[Boolean]("needsSupportAtVenue")).map(booleanTranslator)
-    val gis = adDoc.flatMap(_.getAs[Boolean]("guaranteedInterview")).map(booleanTranslator)
-
-    val fpDoc = doc.getAs[BSONDocument]("civil-service-experience-details")
-    val civilServiceExperienceType = (fpType: CivilServiceExperienceType) => fpDoc.map(
-      _.getAs[CivilServiceExperienceType]("civilServiceExperienceType").contains(fpType)
-    )
-    val internshipTypes = (internshipType: InternshipType) =>
-      fpDoc.map(_.getAs[List[InternshipType]]("internshipTypes").getOrElse(List.empty[InternshipType]).contains(internshipType))
-
-    val civilServant = civilServiceExperienceType(CivilServiceExperienceType.CivilServant).map(booleanTranslator)
-    val fastTrack = civilServiceExperienceType(CivilServiceExperienceType.CivilServantViaFastTrack).map(booleanTranslator)
-    val edip = internshipTypes(InternshipType.EDIP).map(booleanTranslator)
-    val sdipPrevious = internshipTypes(InternshipType.SDIPPreviousYear).map(booleanTranslator)
-    val sdip = internshipTypes(InternshipType.SDIPCurrentYear).map(booleanTranslator)
-    val fastPassCertificate = fpDoc.map(_.getAs[String]("certificateNumber").getOrElse("No"))
-
-    val applicationId = doc.getAs[String]("applicationId").getOrElse("")
-    val progress: ProgressResponse = findProgress(doc, applicationId)
-
-    CandidateProgressReportItem(applicationId, Some(ProgressStatusesReportLabels.progressStatusNameInReports(progress)),
-      schemes.getOrElse(Nil), disability, onlineAdjustments, assessmentCentreAdjustments, gis, civilServant,
-      fastTrack, edip, sdipPrevious, sdip, fastPassCertificate)
-  }
-
-  def toCivilServiceExperienceDetailsReportItem(optDoc: Option[BSONDocument]): Option[CivilServiceExperienceDetailsForDiversityReport] = {
-    optDoc.map { doc =>
-      val civilServiceExperienceType = (fpType: CivilServiceExperienceType) =>
-        doc.getAs[CivilServiceExperienceType]("civilServiceExperienceType").contains(fpType)
-      val internshipTypes = (internshipType: InternshipType) =>
-        doc.getAs[List[InternshipType]]("internshipTypes").getOrElse(List.empty[InternshipType]).contains(internshipType)
-      val civilServant = booleanTranslator(civilServiceExperienceType(CivilServiceExperienceType.CivilServant))
-      val fastTrack = booleanTranslator(civilServiceExperienceType(CivilServiceExperienceType.CivilServantViaFastTrack))
-      val edip = booleanTranslator(internshipTypes(InternshipType.EDIP))
-      val sdipPrevious = booleanTranslator(internshipTypes(InternshipType.SDIPPreviousYear))
-      val sdip = booleanTranslator(internshipTypes(InternshipType.SDIPCurrentYear))
-      val fastPassCertificate = doc.getAs[String]("certificateNumber").getOrElse("No")
-      CivilServiceExperienceDetailsForDiversityReport(Some(civilServant), Some(fastTrack), Some(edip), Some(sdipPrevious),
-        Some(sdip), Some(fastPassCertificate))
-    }
-  }
-
-  def toApplicationForDiversityReport(findProgress: (BSONDocument, String) => ProgressResponse)
-                                         (doc: BSONDocument): ApplicationForDiversityReport = {
-    val schemesDoc = doc.getAs[BSONDocument]("scheme-preferences")
-    val schemes = schemesDoc.flatMap(_.getAs[List[SchemeType]]("schemes"))
-
-    val adDoc = doc.getAs[BSONDocument]("assistance-details")
-    val disability = adDoc.flatMap(_.getAs[String]("hasDisability"))
-    val onlineAdjustments = adDoc.flatMap(_.getAs[Boolean]("needsSupportForOnlineAssessment")).map(booleanTranslator)
-    val assessmentCentreAdjustments = adDoc.flatMap(_.getAs[Boolean]("needsSupportAtVenue")).map(booleanTranslator)
-    val gis = adDoc.flatMap(_.getAs[Boolean]("guaranteedInterview"))
-
-    val civilServiceExperienceDoc = doc.getAs[BSONDocument]("civil-service-experience-details")
-    val civilServiceExperience = toCivilServiceExperienceDetailsReportItem(civilServiceExperienceDoc)
-
-    val applicationId = doc.getAs[String]("applicationId").getOrElse("")
-    val userId = doc.getAs[String]("userId").getOrElse("")
-    val progress: ProgressResponse = findProgress(doc, applicationId)
-
-    ApplicationForDiversityReport(applicationId, userId, Some(ProgressStatusesReportLabels.progressStatusNameInReports(progress)),
-      schemes.getOrElse(List.empty), disability, gis, onlineAdjustments,
-      assessmentCentreAdjustments, civilServiceExperience)
-  }
-
-  def toApplicationForOnlineTestPassMarkReport(findProgress: (BSONDocument, String) => ProgressResponse)
-                                                  (doc: BSONDocument): ApplicationForOnlineTestPassMarkReport = {
-
-    val applicationId = doc.getAs[String]("applicationId").getOrElse("")
-
-    val schemesDoc = doc.getAs[BSONDocument]("scheme-preferences")
-    val schemes = schemesDoc.flatMap(_.getAs[List[SchemeType]]("schemes"))
-
-    val adDoc = doc.getAs[BSONDocument]("assistance-details")
-    val gis = adDoc.flatMap(_.getAs[Boolean]("guaranteedInterview"))
-    val disability = adDoc.flatMap(_.getAs[String]("hasDisability"))
-    val onlineAdjustments = adDoc.flatMap(_.getAs[Boolean]("needsSupportForOnlineAssessment")).map(booleanTranslator)
-    val assessmentCentreAdjustments = adDoc.flatMap(_.getAs[Boolean]("needsSupportAtVenue")).map(booleanTranslator)
-
-    val progress: ProgressResponse = findProgress(doc, applicationId)
-
-    ApplicationForOnlineTestPassMarkReport(
-      applicationId,
-      ProgressStatusesReportLabels.progressStatusNameInReports(progress),
-      schemes.getOrElse(Nil),
-      disability,
-      gis,
-      onlineAdjustments,
-      assessmentCentreAdjustments,
-      toTestResultsForOnlineTestPassMarkReportItem(doc, applicationId))
-  }
-
-  //scalastyle:off method.length
-  private def toTestResultsForOnlineTestPassMarkReportItem(doc: BSONDocument, applicationId: String) = {
-    import config.MicroserviceAppConfig._
-
-    val testGroupsDoc = doc.getAs[BSONDocument]("testGroups")
-
-    val phase1Doc = testGroupsDoc.flatMap(_.getAs[BSONDocument]("PHASE1"))
-    val phase1TestProfile = Phase1TestProfile.bsonHandler.read(phase1Doc.get)
-
-    val situationalScheduleId = cubiksGatewayConfig.phase1Tests.scheduleIds("sjq")
-    val behaviouralScheduleId = cubiksGatewayConfig.phase1Tests.scheduleIds("bq")
-
-    def getTestResult(phase1TestProfile: Phase1TestProfile, scheduleId: Int) = {
-      phase1TestProfile.activeTests.find(_.scheduleId == scheduleId).flatMap { phase1Test =>
-        phase1Test.testResult.map { tr => toTestResult(tr) }
-      }
-    }
-    val behaviouralTestResult = getTestResult(phase1TestProfile, behaviouralScheduleId)
-    val situationalTestResult = getTestResult(phase1TestProfile, situationalScheduleId)
-
-    val phase2DocOpt = testGroupsDoc.flatMap(_.getAs[BSONDocument]("PHASE2"))
-    val etrayTestResult = phase2DocOpt.flatMap { phase2Doc =>
-      val phase2TestProfile = Phase2TestGroup.bsonHandler.read(phase2Doc)
-      phase2TestProfile.activeTests.size match {
-        case 1 => phase2TestProfile.activeTests.head.testResult.map { tr => toTestResult(tr) }
-        case 0 => None
-        case s if s > 1 =>
-          Logger.error(s"There are $s active tests which is invalid for application id [$applicationId]")
-          None
-      }
-    }
-    val reviewedDocOpt = testGroupsDoc.flatMap(_.getAs[BSONDocument]("PHASE3"))
-      .flatMap(_.getAs[BSONArray]("tests")).flatMap(_.getAs[BSONDocument](0))
-      .flatMap(_.getAs[BSONDocument]("callbacks")).flatMap(_.getAs[List[BSONDocument]]("reviewed"))
-
-    val reviewed = reviewedDocOpt.map (_.map(ReviewedCallbackRequest.bsonHandler.read(_)))
-    val latestReviewedOpt = reviewed.map ( _.sortWith { (r1, r2) => r1.received.isAfter(r2.received)}).flatMap(_.headOption)
-
-    val videoInterviewResults = latestReviewedOpt.map { latestReviewed =>
-      VideoInterviewTestResult(
-        toVideoInterviewQuestionTestResult(latestReviewed.latestReviewer.question1),
-        toVideoInterviewQuestionTestResult(latestReviewed.latestReviewer.question2),
-        toVideoInterviewQuestionTestResult(latestReviewed.latestReviewer.question3),
-        toVideoInterviewQuestionTestResult(latestReviewed.latestReviewer.question4),
-        toVideoInterviewQuestionTestResult(latestReviewed.latestReviewer.question5),
-        toVideoInterviewQuestionTestResult(latestReviewed.latestReviewer.question6),
-        toVideoInterviewQuestionTestResult(latestReviewed.latestReviewer.question7),
-        toVideoInterviewQuestionTestResult(latestReviewed.latestReviewer.question8),
-        latestReviewed.calculateTotalScore
-      )
-    }
-    TestResultsForOnlineTestPassMarkReportItem(behaviouralTestResult, situationalTestResult, etrayTestResult, videoInterviewResults)
-  }
-  //scalastyle:on method.length
-
-  private def toVideoInterviewQuestionTestResult(question: ReviewSectionQuestionRequest) = {
-    VideoInterviewQuestionTestResult(
-      question.reviewCriteria1.score,
-      question.reviewCriteria2.score)
-  }
-
-  private def toTestResult(tr: model.persisted.TestResult) = {
-    TestResult(status = tr.status, norm = tr.norm, tScore = tr.tScore, raw = tr.raw, percentile = tr.percentile, sten = tr.sten)
-  }
+trait GeneralApplicationRepoBSONToModelHelper extends BaseBSONToModelHelper {
 
   def toApplicationsForAssessmentAllocation(doc: BSONDocument): ApplicationForAssessmentAllocation = {
     val userId = doc.getAs[String]("userId").get
@@ -267,10 +51,4 @@ trait GeneralApplicationRepoBSONToModelHelper {
     ApplicationForNotification(applicationId, userId, preferredName, applicationStatus)
   }
 
-  private def booleanTranslator(bool: Boolean) = bool match {
-    case true => "Yes"
-    case false => "No"
-  }
-
-  private def extract(key: String)(root: Option[BSONDocument]) = root.flatMap(_.getAs[String](key))
 }
