+++ it/repositories/ApplicationRepositorySpec.scala
@@ -155,109 +155,6 @@ class ApplicationRepositorySpec extends MongoRepositorySpec {
     }
   }
 
-  "Applications report" should {
-    "return an empty list when there are no applications" in {
-      applicationRepo.applicationsReport(frameworkId).futureValue mustBe empty
-    }
-
-    "return a list of non submitted applications when there are only non submitted applications" in {
-      Await.ready({
-        for {
-          _ <- applicationRepo.create("userId1", frameworkId, ApplicationRoute.Faststream)
-          _ <- applicationRepo.create("userId2", frameworkId, ApplicationRoute.Faststream)
-        } yield {
-          Unit
-        }
-      }, timeout)
-
-      val results = applicationRepo.applicationsReport(frameworkId).futureValue
-      results must have size 2
-      results.foreach { case (userId, isNonSubmitted, _) =>
-        isNonSubmitted mustBe true
-        userId must startWith("userId")
-      }
-    }
-
-    "return only submitted applications" in {
-      Await.ready({
-        for {
-          app <- applicationRepo.create("userId1", frameworkId, ApplicationRoute.Faststream)
-          _ <- applicationRepo.addProgressStatusAndUpdateAppStatus(app.applicationId, ProgressStatuses.PREVIEW)
-          _ <- applicationRepo.submit(app.applicationId)
-          app2 <- applicationRepo.create("userId2", frameworkId, ApplicationRoute.Faststream)
-          _ <- applicationRepo.addProgressStatusAndUpdateAppStatus(app2.applicationId, ProgressStatuses.PREVIEW)
-          _ <- applicationRepo.submit(app2.applicationId)
-        } yield {
-          Unit
-        }
-      }, timeout)
-
-      val results = applicationRepo.applicationsReport(frameworkId).futureValue
-      results must have size 2
-      results.foreach { case (_, isNonSubmitted, _) =>
-        isNonSubmitted mustBe false
-      }
-    }
-
-    "return only the applications in a specific framework id" in {
-      Await.ready({
-        for {
-          app <- applicationRepo.create("userId1", frameworkId, ApplicationRoute.Faststream)
-          app2 <- applicationRepo.create("userId2", "otherFramework", ApplicationRoute.Faststream)
-        } yield {
-          Unit
-        }
-      }, timeout)
-
-      val results = applicationRepo.applicationsReport(frameworkId).futureValue
-      results must have size 1
-    }
-
-    "return a list of non submitted applications with submitted applications" in {
-      Await.ready({
-        for {
-          app1 <- applicationRepo.create("userId1", frameworkId, ApplicationRoute.Faststream)
-          _ <- applicationRepo.addProgressStatusAndUpdateAppStatus(app1.applicationId, ProgressStatuses.PREVIEW)
-          _ <- applicationRepo.submit(app1.applicationId)
-          _ <- applicationRepo.create("userId2", frameworkId, ApplicationRoute.Faststream)
-          app3 <- applicationRepo.create("userId3", frameworkId, ApplicationRoute.Faststream)
-          _ <- applicationRepo.addProgressStatusAndUpdateAppStatus(app3.applicationId, ProgressStatuses.PREVIEW)
-          _ <- applicationRepo.submit(app3.applicationId)
-          app4 <- applicationRepo.create("userId4", frameworkId, ApplicationRoute.Faststream)
-          _ <- applicationRepo.addProgressStatusAndUpdateAppStatus(app4.applicationId, ProgressStatuses.PREVIEW)
-          _ <- applicationRepo.submit(app4.applicationId)
-          _ <- applicationRepo.create("userId5", frameworkId, ApplicationRoute.Faststream)
-        } yield {
-          Unit
-        }
-      }, timeout)
-
-      val results = applicationRepo.applicationsReport(frameworkId).futureValue
-      results must have size 5
-      results.filter { case (_, isNonSubmitted, _) => isNonSubmitted } must have size 2
-      results.filter { case (_, isNonSubmitted, _) => !isNonSubmitted } must have size 3
-    }
-  }
-
-  "return the adjustments report" should {
-    "return a list of AdjustmentReports" in {
-      val frameworkId = "FastStream-2016"
-
-      lazy val testData = new TestDataMongoRepository()
-      testData.createApplications(100).futureValue
-
-      val listFut = applicationRepo.adjustmentReport(frameworkId)
-
-      val result = Await.result(listFut, timeout)
-
-      result mustBe a[List[_]]
-      result must not be empty
-      result.head mustBe a[AdjustmentReportItem]
-      result.head.userId must not be empty
-      result.head.applicationId must not be empty
-    }
-  }
-
   "find applications for assessment allocation" should {
     "return an empty list when there are no applications" in {
       lazy val testData = new TestDataMongoRepository()
@@ -421,30 +318,6 @@ class ApplicationRepositorySpec extends MongoRepositorySpec {
         Some("Amber"),
         Some("Fail"),
         Some("Pass"))
-    }
-  }
-
-  "manual assessment centre allocation report" should {
-    "return all candidates that are in awaiting allocation state" in {
-      val testData = new TestDataMongoRepository()
-      testData.createApplications(10, onlyAwaitingAllocation = true).futureValue
-
-      val result = applicationRepo.candidatesAwaitingAllocation(frameworkId).futureValue
-      result must have size 10
-    }
-
-    "not return candidates that are initially awaiting allocation but subsequently withdrawn" in {
-      val testData = new TestDataMongoRepository()
-      testData.createApplications(10, onlyAwaitingAllocation = true).futureValue
-
-      val result = applicationRepo.candidatesAwaitingAllocation(frameworkId).futureValue
-      result.foreach { c =>
-        val appId = applicationRepo.findByUserId(c.userId, frameworkId).futureValue.applicationId
-        applicationRepo.withdraw(appId, WithdrawApplication("testing", None, "Candidate")).futureValue
-      }
-
-      val updatedResult = applicationRepo.candidatesAwaitingAllocation(frameworkId).futureValue
-      updatedResult mustBe empty
     }
   }
