+++ test/controllers/CandidateScoresControllerSpec.scala
@@ -27,8 +27,8 @@ import org.mockito.Mockito._
 import play.api.libs.json.Json
 import play.api.test.Helpers._
 import play.api.test.{ FakeHeaders, FakeRequest, Helpers }
-import repositories.application.{ GeneralApplicationRepository, PersonalDetailsRepository }
-import repositories.{ ApplicationAssessmentRepository, ApplicationAssessmentScoresRepository }
+import repositories.application.GeneralApplicationRepository
+import repositories.{ ApplicationAssessmentRepository, ApplicationAssessmentScoresRepository, PersonalDetailsRepository }
 import testkit.UnitWithAppSpec
 
 import scala.concurrent.Future
