+++ test/controllers/report/OnlineTestPassMarkReportingControllerSpec.scala
@@ -14,16 +14,18 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.report
 
 import config.TestFixtureBase
 import connectors.AuthProviderClient
+import controllers.ReportingController
+import model.report.onlinetestpassmark.{ ApplicationForOnlineTestPassMarkReportItemExamples, TestResultsForOnlineTestPassMarkReportItemExamples }
 import model.report.{ OnlineTestPassMarkReportItem, _ }
 import org.mockito.Matchers._
 import org.mockito.Mockito._
 import play.api.test.Helpers._
 import play.api.test.{ FakeHeaders, FakeRequest, Helpers }
-import repositories.application.GeneralApplicationRepository
+import repositories.application.ReportingRepository
 import repositories.{ ApplicationAssessmentScoresRepository, ContactDetailsRepository, MediaRepository, QuestionnaireRepository, TestReportRepository }
 import testkit.MockitoImplicits.OngoingStubbingExtension
 import testkit.UnitWithAppSpec
@@ -33,8 +35,8 @@ import scala.language.postfixOps
 class OnlineTestPassMarkReportingControllerSpec extends UnitWithAppSpec {
 
   "Online test pass mark report" should {
-    "return nothing if no application exists" in new PassMarkReportTestFixture {
-      when(mockAppRepository.onlineTestPassMarkReport(any())).thenReturnAsync(Nil)
+    "return nothing if no application exists" in new TestFixture {
+      when(mockReportRepository.onlineTestPassMarkReport(any())).thenReturnAsync(Nil)
       when(mockQuestionRepository.findForOnlineTestPassMarkReport).thenReturnAsync(Map.empty)
       when(mockTestResultRepository.getOnlineTestReports).thenReturnAsync(Map.empty)
 
@@ -45,8 +47,8 @@ class OnlineTestPassMarkReportingControllerSpec extends UnitWithAppSpec {
       result mustBe empty
     }
 
-    "return nothing if applications exist, but no questionnaires" in new PassMarkReportTestFixture {
-      when(mockAppRepository.onlineTestPassMarkReport(any())).thenReturnAsync(applications)
+    "return nothing if applications exist, but no questionnaires" in new TestFixture {
+      when(mockReportRepository.onlineTestPassMarkReport(any())).thenReturnAsync(applications)
       when(mockQuestionRepository.findForOnlineTestPassMarkReport).thenReturnAsync(Map.empty)
       when(mockTestResultRepository.getOnlineTestReports).thenReturnAsync(Map.empty)
 
@@ -57,8 +59,8 @@ class OnlineTestPassMarkReportingControllerSpec extends UnitWithAppSpec {
       result mustBe empty
     }
 
-    "return applications and questionnaires if applications and questionnaires exist, but no test results" in new PassMarkReportTestFixture {
-      when(mockAppRepository.onlineTestPassMarkReport(any())).thenReturnAsync(applicationsWithNoTestResults)
+    "return applications and questionnaires if applications and questionnaires exist, but no test results" in new TestFixture {
+      when(mockReportRepository.onlineTestPassMarkReport(any())).thenReturnAsync(applicationsWithNoTestResults)
 
       when(mockQuestionRepository.findForOnlineTestPassMarkReport).thenReturnAsync(questionnairesForNoTestResults)
 
@@ -75,8 +77,8 @@ class OnlineTestPassMarkReportingControllerSpec extends UnitWithAppSpec {
         QuestionnaireReportItemExamples.questionnaire2))
     }
 
-    "return applications with questionnaire and test results" in new PassMarkReportTestFixture {
-      when(mockAppRepository.onlineTestPassMarkReport(any())).thenReturnAsync(applications)
+    "return applications with questionnaire and test results" in new TestFixture {
+      when(mockReportRepository.onlineTestPassMarkReport(any())).thenReturnAsync(applications)
 
       when(mockQuestionRepository.findForOnlineTestPassMarkReport).thenReturnAsync(questionnaires)
       when(mockTestResultRepository.getOnlineTestReports).thenReturnAsync(testResults)
@@ -94,15 +96,15 @@ class OnlineTestPassMarkReportingControllerSpec extends UnitWithAppSpec {
     }
   }
 
-  trait PassMarkReportTestFixture extends TestFixtureBase {
+  trait TestFixture extends TestFixtureBase {
     val frameworkId = "FastStream-2016"
 
-    val mockAppRepository = mock[GeneralApplicationRepository]
+    val mockReportRepository = mock[ReportingRepository]
     val mockQuestionRepository = mock[QuestionnaireRepository]
     val mockTestResultRepository = mock[TestReportRepository]
     val mockMediaRepository = mock[MediaRepository]
     val controller = new ReportingController {
-      val appRepository = mockAppRepository
+      val reportRepository = mockReportRepository
       val cdRepository = mock[ContactDetailsRepository]
       val authProviderClient = mock[AuthProviderClient]
       val questionnaireRepository = mockQuestionRepository
