+++ test/controllers/report/ReportingControllerSpec.scala
@@ -14,13 +14,14 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.report
 
 import config.TestFixtureBase
 import connectors.AuthProviderClient
 import connectors.ExchangeObjects.Candidate
+import controllers.ReportingController
 import mocks._
-import mocks.application.DocumentRootInMemoryRepository
+import mocks.application.ReportingInMemoryRepository
 import model.Address
 import model.PersistedObjects.ContactDetailsWithId
 import model.report._
@@ -29,7 +30,7 @@ import org.mockito.Mockito._
 import play.api.libs.json.JsArray
 import play.api.test.Helpers._
 import play.api.test.{ FakeHeaders, FakeRequest, Helpers }
-import repositories.application.GeneralApplicationRepository
+import repositories.application.ReportingRepository
 import repositories.{ ApplicationAssessmentScoresRepository, ContactDetailsRepository, MediaRepository, QuestionnaireRepository, TestReportRepository }
 import testkit.UnitWithAppSpec
 
@@ -39,7 +40,7 @@ import scala.language.postfixOps
 class ReportingControllerSpec extends UnitWithAppSpec {
 
   class TestableReportingController extends ReportingController {
-    override val appRepository: GeneralApplicationRepository = DocumentRootInMemoryRepository
+    override val reportRepository: ReportingRepository = ReportingInMemoryRepository
     override val cdRepository: ContactDetailsRepository = new ContactDetailsInMemoryRepository
     override val authProviderClient: AuthProviderClient = mock[AuthProviderClient]
       override val questionnaireRepository: QuestionnaireRepository = QuestionnaireInMemoryRepository
@@ -87,7 +88,7 @@ class ReportingControllerSpec extends UnitWithAppSpec {
 
     "return no adjustments if there's no data on the server" in new TestFixture {
       val controller = new TestableReportingController {
-        override val appRepository = new DocumentRootInMemoryRepository {
+        override val reportRepository = new ReportingInMemoryRepository {
           override def adjustmentReport(frameworkId: String): Future[List[AdjustmentReportItem]] = {
             Future.successful(List.empty[AdjustmentReportItem])
           }
