+++ test/mocks/application/DocumentRootInMemoryRepository.scala
@@ -93,45 +93,7 @@ class DocumentRootInMemoryRepository extends GeneralApplicationRepository {
     lastName: Option[String], dateOfBirth: Option[LocalDate],
     userIds: List[String] = List.empty): Future[List[Candidate]] = Future.successful(List.empty[Candidate])
 
-  override def applicationsReport(frameworkId: String): Future[List[(String, IsNonSubmitted, PreferencesWithContactDetails)]] = {
-    val app1 = ("user1", true, PreferencesWithContactDetails(Some("John"), Some("Smith"), Some("Jo"),
-      Some("user1@email.com"), None, Some("location1"), Some("location1Scheme1"), Some("location1Scheme2"), Some("location2"),
-      Some("location2Scheme1"), Some("location2Scheme2"), None, Some("2016-12-25 13:00:14")))
-    val app2 = ("user2", true, PreferencesWithContactDetails(Some("John"), Some("Smith"), Some("Jo"),
-      None, None, None, None, None, None, None, None, None, None))
-    Future.successful(app1 :: app2 :: Nil)
-  }
-
   override def findCandidateByUserId(userId: String): Future[Option[Candidate]] = Future.successful(None)
-
-  override def candidateProgressReportNotWithdrawn(frameworkId: String): Future[List[CandidateProgressReportItem]] =
-    candidateProgressReport(frameworkId)
-
-  override def candidateProgressReport(frameworkId: String): Future[List[CandidateProgressReportItem]] = Future.successful(List(
-    CandidateProgressReportItem("", Some("registered"),
-      List(SchemeType.DigitalAndTechnology, SchemeType.Commercial), None, None, None, None, None, None, None, None, None, None))
-  )
-
-  override def diversityReport(frameworkId: String): Future[List[ApplicationForDiversityReport]] = ???
-
-  override def onlineTestPassMarkReport(frameworkId: String): Future[List[ApplicationForOnlineTestPassMarkReport]] = ???
-
-  override def adjustmentReport(frameworkId: String): Future[List[AdjustmentReportItem]] =
-    Future.successful(
-      List(
-        AdjustmentReportItem("1", Some("11"), Some("John"), Some("Smith"), Some("Spiderman"), None, None, Some("Yes"),
-          Some(ApplicationStatus.SUBMITTED), Some("Need help for online tests"), Some("Need help at the venue"),
-          Some("Yes"), Some("A wooden leg"),
-          Some(Adjustments(Some(List("etrayTimeExtension")),Some(true),Some(AdjustmentDetail(Some(55),None,None)),None)),None),
-        AdjustmentReportItem("2", Some("22"), Some("Jones"), Some("Batman"), None, None, None, None,
-          Some(ApplicationStatus.PHASE1_TESTS), None, Some("Need help at the venue"), None, None,
-          Some(Adjustments(Some(List("etrayTimeExtension")),Some(true),Some(AdjustmentDetail(Some(55),None,None)),None)),None),
-        AdjustmentReportItem("3", Some("33"), Some("Kathrine"), Some("Jones"), Some("Supergirl"), None, None, None,
-          Some(ApplicationStatus.PHASE1_TESTS_PASSED), Some("Need help for online tests"), None,
-          Some("Yes"), Some("A glass eye"),
-          Some(Adjustments(Some(List("etrayTimeExtension")),Some(true),Some(AdjustmentDetail(Some(55),None,None)),None)),None)
-      )
-    )
 
   override def findApplicationIdsByLocation(location: String): Future[List[String]] = Future.successful(List())
 
@@ -151,30 +113,17 @@ class DocumentRootInMemoryRepository extends GeneralApplicationRepository {
 
   override def rejectAdjustment(applicationId: String): Future[Unit] = Future.successful(Unit)
 
-  override def candidatesAwaitingAllocation(frameworkId: String): Future[List[CandidateAwaitingAllocation]] =
-    Future.successful(
-      List(
-        CandidateAwaitingAllocation("1", "John", "Smith", "Spiderman", "London", Some("Some adjustments"), new LocalDate(1988, 1, 21)),
-        CandidateAwaitingAllocation("2", "James", "Jones", "Batman", "Bournemouth", Some("Some adjustments"), new LocalDate(1992, 11, 30)),
-        CandidateAwaitingAllocation("3", "Katherine", "Jones", "Supergirl", "Queer Camel", None, new LocalDate(1990, 2, 12))
-      )
-    )
-
   override def findFailedTestForNotification(failedTestType: FailedTestType): Future[Option[NotificationFailedTest]] = {
     Future.successful(Some(NotificationFailedTest("31009ccc-1ac3-4d55-9c53-1908a13dc5e1", "fbb466a3-13a3-4dd0-93d6-9dfa764a5555", "George")))
   }
 
   override def gisByApplication(applicationId: String): Future[Boolean] = ???
 
-  override def overallReportNotWithdrawnWithPersonalDetails(frameworkId: String): Future[List[ReportWithPersonalDetails]] = ???
-
   override def allocationExpireDateByApplicationId(applicationId: String): Future[Option[LocalDate]] = ???
 
   override def updateStatus(applicationId: String, applicationStatus: ApplicationStatus): Future[Unit] = ???
 
   override def applicationsWithAssessmentScoresAccepted(frameworkId: String): Future[List[ApplicationPreferences]] = ???
-
-  override def allApplicationAndUserIds(frameworkId: String): Future[List[PersonalDetailsAdded]] = ???
 
   override def nextAssessmentCentrePassedOrFailedApplication(): Future[Option[ApplicationForNotification]] = ???
