+++ test/services/generaldetails/CandidateDetailsServiceSpec.scala
@@ -22,7 +22,7 @@ import model.persisted.ContactDetailsExamples._
 import model.persisted.PersonalDetailsExamples._
 import org.mockito.Matchers.{ eq => eqTo, _ }
 import org.mockito.Mockito._
-import repositories.CivilServiceExperienceDetailsRepository
+import repositories.civilserviceexperiencedetails.CivilServiceExperienceDetailsRepository
 import repositories.contactdetails.ContactDetailsRepository
 import repositories.personaldetails.PersonalDetailsRepository
 import services.{ AuditService, BaseServiceSpec }
